Fix bugs with reset and plugin loading.
Modules are no longer loaded into a global array; instead, now
they get freshly loaded on each call to the plugin load function.
If a plugin was already loaded and silent=False, an error is thrown
(because this means you have imported a module that conflicts with
a plugin name, which is bad), but if silent=True (i.e. we've already
successfully ran this function once) the plugin loader reloads all
the plugins.